requirements.txt
torch
torchvision
timm
opencv-python
numpy
pandas
scikit-learn
pyyaml
tqdm
matplotlib
seaborn
gdown
wandb